v3 refractor done, move to ML...

+++ watchdog.py
@@ -26,52 +26,71 @@ word = word + str(int(word[add_position]) + add_constant)
 
 device_name = params['device_info']['device_name']
 
-# set the error count for each exchange to 0
-# ###PAUL this is currently implemented wrong. would be better to do params['systemd_control']['active_services].keys()
-error_count_dict = {'trade': {}, 'price': {}, }
-# for data_type in params['systemd_control']['active_services''.keys():
-#     for params['systemd_control']\
-for ex in params['active_services']['trades'].keys():
-    error_count_dict['trade'][ex] = 0
-    error_count_dict['price'][ex] = 0
-
-
-
 
 # used a few times to restart services, good because it allows for mode switching easily for manual updates
-def restart_service(service, script, pword=word, mode='hard_restart'):
-    send_email(subject='WATCHDOG UPDATE: ' + service + ' ---- restart',
-               message=('restarting: \n'
-                        + '---- service: ' + service
-                        + '---- script: ' + script),
-               script=script)
+def restart_service(service_type, params_t, exchange=None, port=None, pword=word, mode='hard_restart'):
+    """takes given parameters and will restart (if cooldown doesn't override the restart)
 
-    if mode == 'restart':
-        command = 'sudo systemctl restart ' + service + '.service'
-        p = os.system('echo %s|sudo -S %s' % (pword, command))
+    output:
+        params (dict): NOTE, this fn takes params in its own space, modifies it in functions that treat it globally
+    """
+    # can only receive an exchange OR a portfolio...
+    if exchange is not None and port is not None:
+        print('can only take a port OR an exchange')
+        raise IOError
+    if exchange is not None:  # we are restarting a trade collector or price maker
+        _ = exchange
+        service_dict = params_t['active_services'][service_type][exchange]
+    if port is not None:  # restarting a portfolio
+        _ = port
+        service_dict = params_t['active_services'][service_type][port]
 
-    if mode == 'hard_restart':
-        command = 'sudo chmod u+x ' + script \
-                  + '&& sudo systemctl stop ' + service + '.service ' \
-                  + '&& sudo systemctl daemon-reload' \
-                  + '&& sudo systemctl enable ' + service + '.service' \
-                  + '&& sudo systemctl restart ' + service + '.service'
+    service = service_dict['service']
+    script = service_dict['script']
+    last_restart = service_dict['last_restart']
+    cool_down = service_dict['cool_down']
 
-        p = os.system('echo %s|sudo -S %s' % (pword, command))
+    # hasn't been long enough... return the same params_t dict and wait till next iteration
+    if last_restart + cool_down < time.time():
+        cooldown_str = ('---- WATCHDOG: COOLDOWN prevented restart ---- service_type ' + service_type + ' ---- ' + _
+                        + ' --- aka script ---- ' + script + ' \n') * 5
+        print(cooldown_str)
+        return params_t
 
-    return None
+    else:
+        send_email(subject='WATCHDOG UPDATE: ' + service + ' ---- restart',
+                   message=('restarting: \n'
+                            + '---- service: ' + service
+                            + '---- script: ' + script),
+                   script=script)
+        if mode == 'restart':
+            command = 'sudo systemctl restart ' + service + '.service'
+            p = os.system('echo %s|sudo -S %s' % (pword, command))
+        if mode == 'hard_restart':
+            command = 'sudo chmod u+x ' + script \
+                      + '&& sudo systemctl stop ' + service + '.service ' \
+                      + '&& sudo systemctl daemon-reload' \
+                      + '&& sudo systemctl enable ' + service + '.service' \
+                      + '&& sudo systemctl restart ' + service + '.service'
+            p = os.system('echo %s|sudo -S %s' % (pword, command))
+
+        # reset last_restart and put the new service_dict back into params_t
+        service_dict['last_restart'] = time.time()
+        params_t['active_services'][service_type][_] = service_dict
+
+        return params_t
 
 
 def trade_watchdog():
     """checks if we are getting trades and will hard reset the trade datascrape service
     """
 
+    global params
+
     print("-=-=-=-=-=-=-=-= ALGOS WATCHDOG: ------------ CHECKING TRADES -=-=-=-=-=-=-=-=\n", flush=True)
 
     for exchange in params['active_services']['trades'].keys():
-        # keep this here, bottom line needed for top
         service = params['active_services']['trades'][exchange]['service']
-        script = params['active_services']['trades'][exchange]['script']
         restart_notification_string = 3 * ('NOT GETTING TRADES - exchange: ' + exchange +
                                            ' - RESTARTING  -  ' + service + '\n')
 
@@ -84,21 +103,29 @@ def trade_watchdog():
             if time_since_last_btc_trade > no_trade_time:  # restart systemd service
                 # restart it
                 print(restart_notification_string, flush=True)
-                restart_service(service, script)
+                params = restart_service(service_type='trades',
+                                         params_t=params,
+                                         exchange=exchange,
+                                         )
 
             else:
-                error_count_dict['trade'][exchange] = 0
+                params['active_services']['trades'][exchange]['error_count'] = 0
                 print("-=-=-=-=-=-=-=-= ALGOS WATCHDOG: trades being received for     " + exchange + "    -=-=-=-=-=\n",
                       flush=True)
 
         # some exchanges are slow, so we will give 3 checks untill we force a restart
         except (FileNotFoundError, IndexError):
-            error_count_dict['trade'][exchange] += 1
-            if error_count_dict['trade'][exchange] > 3:
+            params['active_services']['trades'][exchange]['error_count'] += 1
+            if params['active_services']['trades'][exchange]['error_count'] > 3:
                 # restart it
                 print('-=-=-=-=-=-=-=-= ALGOS WATCHDOG: max ERROR count for exchange: ' + exchange, flush=True)
                 print(restart_notification_string, flush=True)
-                restart_service(service, script)
+                # restart_service(service, script)
+                params = restart_service(service_type='trades',
+                                         params_t=params,
+                                         exchange=exchange,
+                                         port=None,
+                                         )
 
     return None
 
@@ -107,10 +134,13 @@ def price_crypto_making_watchdog():
     """checks if we are making prices from trades and will hard reset orders.py service
     """
 
-    global error_count_dict
+    global params
 
     print('-=-=-=-=-=-=-=-= ALGOS WATCHDOG: ------------ CHECKING PRICES -=-=-=-=-=-=-=-=\n', flush=True)
     for exchange in params['active_services']['prices'].keys():
+        service = params['active_services']['prices'][exchange]['service']
+        service_restart_str = ('-=-=-=-=- RESTARTING CRYPTO PRICE MAKER: ' + service + ' -=-=-=-=-=-=-') * 5
+
         try:
             check_ticker = params['active_services']['trades'][exchange]['ticker_to_check']
             prices = get_data(data_type='prices', ticker=check_ticker, exchange=exchange)
@@ -121,33 +151,28 @@ def price_crypto_making_watchdog():
             seconds_since_last_price = time.time() - epoch_price_time
 
             if seconds_since_last_price > params['active_services']['prices'][exchange]['no_price_time']:
-                service = params['active_services']['prices']['crypto']['service']
-                script = params['active_services']['prices']['crypto']['script']
-                print(2 * '-=-=-=-=- RESTARTING CRYPTO PRICE MAKER: ' + service + ' -=-=-=-=-=-=-',
-                      flush=True)
-
                 # restart it
-                restart_service(service, script)
+                print(service_restart_str, flush=True)
+                params = restart_service(service_type='prices',
+                                         params_t=params,
+                                         exchange=exchange,
+                                         )
 
             else:
                 print('-=-=-=-=-=-=-=-= ALGOS WATCHDOG prices being made for    ' + exchange + '    -=-=-=-=-=-=-=-=\n',
                       flush=True)
-                error_count_dict['price'][exchange] = 0
+                params['active_services']['prices'][exchange]['error_count'] = 0
 
         except FileNotFoundError:  # sometimes the price may not have been made ... weird stuff happens...
-            error_count_dict['price'][exchange] += 1
+            params['active_services']['prices'][exchange]['error_count'] += 1
 
-            if error_count_dict['price'][exchange] > 3:
-                # ###PAUL_debug i dont think this is the right way to handle this but its late
-                print('-=-=-=-=-=- ALGOS WATCHDOG: price making ERROR  -=-=-=-=-=-', flush=True)
-                # raise RuntimeError
-                service = params['active_services']['prices']['crypto']['service']
-                script = params['active_services']['prices']['crypto']['script']
-                print(2 * '-=-=-=-=- ALGOS - RESTARTING CRYPTO PRICE MAKER: ' + service + ' -=-=-=-=-=-=-',
-                      flush=True)
-
-                # restart service
-                restart_service(service, script)
+            if params['active_services']['prices'][exchange]['error_count'] > 3:
+                print('-=-=-=-=-=- ALGOS WATCHDOG: price making RESTART FROM ERRORS  -=-=-=-=-=-' * 10, flush=True)
+                print(restart_str, flush=True)
+                params = restart_service(service_type='prices',
+                                         params_t=params,
+                                         exchange=exchange,
+                                         )
 
     return None
 
@@ -158,6 +183,8 @@ def check_if_orders_being_updated():
     if True, then orders are being updated then we are good.
     """
 
+    global params
+
     print('-=-=-=-=-=-=-=-= ALGOS WATCHDOG: ------------ CHECKING LIVE BOTS -=-=-=-=-=-=-=-= on ---- ' \
           + device_name + '\n', flush=True)
 
@@ -166,42 +193,50 @@ def check_if_orders_being_updated():
     if len(active_ports) == 0:
         print('-=-=-=-=-=-=-=-= no bots running on ---- ' + device_name, flush=True)
     for port_name in active_ports:
-        exchange = params['active_services']['ports'][port_name]['exchange']
-        service = params['active_services']['ports'][port_name]['service']
-        script = params['active_services']['ports'][port_name]['script']
-
         try:
+            exchange = params['active_services']['ports'][port_name]['exchange']
+            service = params['active_services']['ports'][port_name]['service']
+            restart_str = (2 * '-=-=-=-=- ALGOS - RESTARTING LIVE BOT: ' + service + ' -=-=-=-=-=-=-')*5
+
             fp = get_data_file_path(data_type='last_order_check',
                                     ticker=None,
                                     date='live',
                                     port=port_name,
                                     exchange=exchange)
-        except FileNotFoundError:
-            print(2 * '\n ---- ALGOS - LIVEBOT CHECK: order file not found  --> \n' + str(fp), flush=True)
-            print('---- PROBLEM: last_order_check not found for port ---- ' + port_name, flush=True)
-            return None
 
-        with open(fp, 'r') as f:
-            last_update_time = f.readline()
-        os.chmod(fp, 0o777)
+            with open(fp, 'r') as f:
+                last_update_time = f.readline()
+            os.chmod(fp, 0o777)
 
-        last_update_time = float(last_update_time)
-        time_since_last_order_update = time.time() - last_update_time
+            last_update_time = float(last_update_time)
+            time_since_last_order_update = time.time() - last_update_time
 
-        if time_since_last_order_update > 60:
-            print(2 * '-=-=-=-=- ALGOS - RESTARTING LIVE BOT: ' + service + ' -=-=-=-=-=-=-',
-                  flush=True)
+            if time_since_last_order_update > 60:
+                # restart it
+                print(restart_str, flush=True)
+                params = restart_service(service_type='ports',
+                                         params_t=params,
+                                         port=port_name,
+                                         )
 
-            # restart it
-            restart_service(service, script)
+            else:  # orders being checked / placed for this portfolio
+                params['active_services']['ports'][port_name]['error_count'] = 0
+                print('-=-=-=-=-=-=-=-= ALGOS WATCHDOG:   port   ' + port_name + '  ---- is active on   ' + exchange
+                      + '   -=-=-=-=-=\n',
+                      flush=True)
+        except:
+            params['active_services']['ports'][port_name]['error_count'] += 1
 
-        else:  # orders being checked / placed for this portfolio
-            print('-=-=-=-=-=-=-=-= ALGOS WATCHDOG:   port   ' + port_name + '  ---- is active on   ' + exchange
-                  + '   -=-=-=-=-=\n',
-                  flush=True)
-
+            if params['active_services']['ports'][port_name]['error_count'] > 3:
+                print('---- WACHDOG: restarting port due to ERRORS ----'*5, flush=True)
+                print(restart_str, flush=True)
+                params = restart_service(service_type='ports',
+                                         params_t=params,
+                                         port=port_name,
+                                         )
 
 print('---- WATCHDOG: started ---- \n' * 10, flush=True)
+
 
 # trade reception watchdog
 check_trades_interval = params['active_services']['watchdog']['check_trades_interval']
